pyproject.toml
[project]
name = "reward-fm"
version = "0.0.1"
description = "PyTorch implementation of a Reward Foundation Model (RFM)"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Anthony Liang", email = "aliang80@usc.edu" }
]

dependencies = [
    # Core ML dependencies (all work on macOS)
    "torch>=2.4.0",
    "torchvision",
    "transformers==4.53.2",
    "accelerate>=1.9.0",
    "einops>=0.8.0",
    "peft",
    "safetensors>=0.6.0rc0",
    "sentence-transformers>=2.0.0",
    
    # Data & processing
    "datasets>=3.6.0",
    "h5py>=3.10.0",
    "imageio>=2.37.0",
    "numpy>=1.21.0",
    "pillow>=9.0.0",
    "opencv-python-headless>=4.8.0",
    
    # HuggingFace & VLM
    "huggingface-hub[hf-xet]>=0.33.0",
    "qwen-vl-utils>=0.0.11",
    "google-generativeai>=0.8.0",
    
    # Server & API
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "requests>=2.32.4",
    "aiohttp>=3.9.0",
    
    # Config & utilities
    "pyyaml>=6.0.0",
    "omegaconf>=2.3.0",
    "pyrallis>=0.3.0",
    "rich>=14.0.0",
    "tqdm>=4.67.1",
    "codetiming>=1.4.0",
    "scipy>=1.15.3",
]

[project.optional-dependencies]
# Only bitsandbytes is optional (doesn't work on macOS ARM64)
gpu = [
    "bitsandbytes",
]

# Development extras
dev = [
    "evaluate>=0.4.3",
    "matplotlib>=3.5.0",
    "wandb>=0.20.1",
    "soundfile>=0.13.1",
    "ipdb>=0.13.13",
    "hf_transfer>=0.1.0",
    "gradio>=4.0.0",
    "ffmpeg",
    "termcolor>=3.1.0",
    "robosuite==1.4.0",
    "opencv-python-headless>=4.8.0",
    "opencv-contrib-python-headless>=4.8.0", # Additional OpenCV modules
    "einops>=0.8.0",
    "num2words",
    "seaborn",
    "moviepy",
    "wandb[media]",
    "torchcodec",
    "decord>=0.6.0",
]

[project.urls]
Repository = "https://github.com/aliang8/reward_fm"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["rfm"]


# ==================================================================================
# Tool Configuration
# Settings for various development tools like 'uv', 'ruff' (linter), etc.
# ==================================================================================

# Configuration for the 'uv' package manager
[tool.uv]

# Tool configuration for the 'ruff' linter and formatter
[tool.ruff]
line-length = 120
target-version = "py310"
fix = true
unsafe-fixes = true

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "RUF"]
ignore = ["T201"]

[tool.ruff.lint.isort]
combine-as-imports = true
split-on-trailing-comma = true

[tool.ruff.format]
preview = true
